requirements.txt
eventlet==0.33.0
requests==2.26.0
scalecodec==1.2.4
substrate_interface==1.7.1
behave==1.2.6
web3==6.5.0
pytest==7.4.0
python-on-whales==0.64.0
peaq-py==0.0.12
